don't download if not defined

+++ downloadExperiment.sh
@@ -1,7 +1,9 @@
-echo "Downloading Experiment\n"
-rm -r /root/experiment/
-git clone https://$gitUsername:$gitPassword@$gitURL /root/experiment/
-if [[ $experimentVersion != "latest" ]]; then
-	git checkout $experimentVersion
+if [[ $experimentVersion != "" ]]; then
+	echo "Downloading Experiment\n"
+	rm -r /root/experiment/
+	git clone https://$gitUsername:$gitPassword@$gitURL /root/experiment/
+	if [[ $experimentVersion != "latest" ]]; then
+		git checkout $experimentVersion
+	fi
+	echo "Experiment Downloaded\n"
 fi
-echo "Experiment Downloaded\n"
